+++ Source/Application/Assets/CustomECCWarning.html
@@ -10,4 +10,13 @@
 
 <p>If you find any faults in the source code of either EllipticCurveKit, <em><a href="https://github.com/OpenZesame/Zesame">Zesame (Zilliqa Swift SDK)</a></em> or <em>Zhip</em>, please report issues on <em>Github</em> and/or fix the bug and submit a pull request.</p>
 
-<p><strong>By pressing the button <em>"Accept"</em> below, you accept the risks associated with using this wallet and its unreliable <em>ECC</em> and you free all and any author who has developed <em>Zhip, Zesame, EllipticCurveKit</em> from all liability.</strong></p>
+<p><strong>By pressing the button <em>"Accept"</em> below, you accept the risks associated with using this wallet and its unreliable <em>ECC</em> and that the use of this wallet falls under the license of the software which states:</strong></p>
+<p><strong>
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR<br>
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,<br>
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE<br>
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER<br>
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,<br>
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE<br>
+SOFTWARE.
+</strong></p>
