+++ Source/Scenes/0_Onboarding/2_AskForCrashReportingPermissions/AskForCrashReportingPermissionsView.swift
@@ -84,7 +84,7 @@ private extension AskForCrashReportingPermissionsView {
         }
 
         disclaimerTextView.withStyle(.nonEditable) {
-            $0.text(€.Text.disclaimer)
+            $0.text(€.Text.disclaimer).isSelectable(false)
         }
 
         hasReadDisclaimerCheckbox.withStyle(.default) {
