+++ Source/Scenes/0_Onboarding/4_WarningCustomECC/WarningCustomECCViewModel.swift
@@ -41,11 +41,11 @@ final class WarningCustomECCViewModel: BaseViewModel<
 > {
 
     private let useCase: OnboardingUseCase
-    private let isDismissable: Bool
+    private let isDismissible: Bool
 
-    init(useCase: OnboardingUseCase, isDismissable: Bool) {
+    init(useCase: OnboardingUseCase, isDismissible: Bool) {
         self.useCase = useCase
-        self.isDismissable = isDismissable
+        self.isDismissible = isDismissible
     }
 
     // swiftlint:disable:next function_body_length
@@ -56,7 +56,7 @@ final class WarningCustomECCViewModel: BaseViewModel<
 
         let isAcceptButtonEnabled = input.fromView.didScrollToBottom.map { true }
 
-        if isDismissable {
+        if isDismissible {
             input.fromController.rightBarButtonContentSubject.onBarButton(.done)
             input.fromController.rightBarButtonTrigger
                 .do(onNext: { userDid(.dismiss) })
@@ -71,7 +71,7 @@ final class WarningCustomECCViewModel: BaseViewModel<
         ]
 
         return Output(
-            isAcceptButtonVisible: Driver.just(!isDismissable),
+            isAcceptButtonVisible: Driver.just(!isDismissible),
             isAcceptButtonEnabled: isAcceptButtonEnabled
         )
     }
